+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed Fluid/Ore Laser drill JEI recipe closes #1467
 * Fixed Fluid Laser Drill recipe requiring nether dimension and specific biomes
 * Fixed manual closes #1466
+* Changed Recipes to use wooden chests only
 
 # Version 3.6.5